Increase buffer size of REST API POST body
The maximum length of a Moodle username is 100 characters, the maximum length of a password is 255 characters and the maximum length of a service shortname is 255 characters. The length of the string, "username=&password=&service=" is 28 characters. So the maximum length of the HTTP POST body is 100 + 255 + 255 + 28 = 638. An extra byte is allocated to store the terminating null.

I may need to revise this buffer's size again, as I have seen references in Moodle's source about allowing users to login using their email as well. Also, the body is not URL encoded at the moment, which should be addressed.