Restructure if statements
HTTP status lines are reasonably short, so a single recv() should get enough bytes for a single round of parsing to succeed, hence why a check for a successful return value is done first. For the same reason as stated previously, the next likely return value from the parse function call is likely to be an error. As the parse function can only return success, error or error again values, it is sufficient to check if the return value is not equal to the error again value, which also serves as defensive programming against an erroneous return value being introduced. The defensive programming referred to prior is also part of the reason why no check needs to be performed for if the return value was error again and why it may assumed that if execution has gotten this far, that that is the value of the return value. The other part of the reason is that we have a return statement in the if statement for errors.

Finally, the buffer for processing the status line should be large enough to hold everything up to the status code part of the status line. If the buffer runs out of space at this point of processing the HTTP response, then the reason phrase part of the HTTP status line was too long for the program to handle.